environmental object and keyboard focuse
Focused on the keyboard for the responses needed by the user. Also connected the environmental variable to be able to be shared with all views. Now need to connect the other variables that need to be environmental as well as after submit or click home everything all views should go back to beginning as well as have the info blank